Add files via upload
In many aspects of research, missing data is frequently common. What do
we mean by missing data, well missing data is as intuitive as it sounds. Any
data that was intended to be observed but hasn’t is classified as missing data.
Researchers have relied on a variety of ad-hoc methods for decades in an effort
to either complete the data by eliminating invalid cases (cases with any missing
data) or adding predictions for the missing values so the data analysis could
be carried out. However, these methods have their downsides, albeit they are
convenient due to the methods having to rely on assumptions on how the missing
data came about. In this report, we will discuss the problem that missing data
brings and the different types of missing data. Then some of these methods that
have been used originally will be discussed, looking at some results of using these
methods and then discussing why these methods don’t actually work. Then one
method, multiple imputation will be discussed which is a widely popular method
that was developed to tackle missing data by creating feasible replacements so
that a non-bias analysis can take place.